refactor(rust-api): extract project + asset handlers into modules (sc-1636)
Move the project endpoint handlers (list/create/get project,
reindex_project_endpoint) into projects.rs and the asset handlers
(list/get/import asset, update_asset_status, delete/purge asset, plus
the import upload helper) into assets.rs. health/access/verify stay at
the root as core endpoints.

Same pattern as the prior handler splits: `use super::*;` + pub(crate)
handlers glob-re-imported at the root. Behavior-preserving.
lib.rs 7,643 -> 7,471 lines.
